// src/library/types/Light.tsx
export interface LightProps{
    $height?: string;
    $width?: string;
    $borderWidth?: number;
    $rounded?: boolean;
    $speed?: string;
    $contentColor?: string;
    $color1?: string;
    $color2?: string;
    $color3?: string;
    $fontsize?:string;
    $fontColor?:string;
    $textAlign?:string;
    $outline?:boolean;
    $margin?:string;
  };